Almost done with Combat Core
Added Collision Relay system.  Finished up Extension and Controller.
Wrapped up Destructable object base script.  Finished most of Entity.
Small adjustments across the board to make things better/simpler.

+++ Assets/Scripts/Combat Core/StatusComponent.cs
@@ -58,7 +58,7 @@ public class StatusComponent : ISerializable
 	public virtual void OnRooted(Entity subject){ }
 
 	// Called when the subject is healed
-	public virtual void OnHealed(Entity subject, int healAmount){ }
+	public virtual void OnHealed(Entity subject, float healAmount){ }
 
 	// For serialization
 	public void GetObjectData(SerializationInfo info, StreamingContext context)
